Added image and contributing section
This commit is based on some of the things learnt in the campus-experts technical writing workshop. An image has been added to make the readme more visual, the description provides context for the repo and there has been a contributing guideline added. Build and test instructions we're simple and straightforward enough i believe so they remained the same